requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
coverage==6.3.2
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.0.3
django-admin-searchable-dropdown==1.2
django-cors-headers==3.11.0
django-countries==7.3.2
django-extensions==3.1.5
django-import-export==2.9.0
django-jazzmin==2.5.0
django-shell-plus==1.1.7
django-showurls==1.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
et-xmlfile==1.1.0
future==0.18.2
gunicorn==20.1.0
iniconfig==1.1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.7
pytest==7.0.1
pytest-cov==3.0.0
pytest-django==4.5.2
python-dotenv==0.20.0
python-environ==0.4.54
pytz==2022.1
PyYAML==6.0
serializers==0.2.4
sqlparse==0.4.2
tablib==3.2.1
tomli==2.0.1
typing_extensions==4.1.1
tzdata==2022.1
uuid==1.30
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
